+++ tests/test_machine.py
@@ -25,11 +25,12 @@ class RunTestCase:
         super().setUpClass()
         cls.cls_exit_stack = contextlib.ExitStack()
         cls.moncic = make_moncic()
-        cls.images = cls.cls_exit_stack.enter_context(cls.moncic.images())
+        cls.session = cls.cls_exit_stack.enter_context(cls.moncic.session())
 
     @classmethod
     def tearDownClass(cls):
-        cls.cls_exit_stack.close()
+        with privs.root():
+            cls.cls_exit_stack.close()
         cls.images = None
         cls.moncic = None
         cls.cls_exit_stack = None
@@ -38,9 +39,9 @@ class RunTestCase:
     @contextlib.contextmanager
     def system(self) -> Generator[System, None, None]:
         with privs.root():
-            if self.distro_name not in self.images.list_images():
+            if self.distro_name not in self.session.images().list_images():
                 raise unittest.SkipTest(f"Image {self.distro_name} not available")
-            with self.images.system(self.distro_name) as system:
+            with self.session.images().system(self.distro_name) as system:
                 if not os.path.exists(system.path):
                     raise unittest.SkipTest(f"Image {self.distro_name} has not been bootstrapped")
                 yield system
@@ -178,34 +179,46 @@ class RunTestCase:
                 self.assertEqual(res.stderr, b"")
 
                 binds = list(container.binds())
-                self.assertEqual(len(binds), 1)
+                self.assertIn(len(binds), (1, 2))
                 self.assertEqual(binds[0].source, workdir)
                 self.assertEqual(binds[0].destination, "/media/" + os.path.basename(workdir))
                 self.assertEqual(binds[0].bind_type, "rw")
                 self.assertEqual(binds[0].mount_options, [])
+                if len(binds) == 2:
+                    # self.assertEqual(binds[1].source, workdir)
+                    self.assertEqual(binds[1].destination, "/var/cache/apt/archives")
+                    self.assertEqual(binds[1].bind_type, "rw")
+                    self.assertEqual(binds[1].mount_options, [])
 
     def test_bind_mount_rw(self):
         with tempfile.TemporaryDirectory() as workdir:
             # By default, things are run as root
             container_config = ContainerConfig()
-            container_config.binds.append(BindConfig(source=workdir, destination="/media/workdir", bind_type="rw"))
+            container_config.binds.append(
+                    BindConfig.create(source=workdir, destination="/media/workdir", bind_type="rw"))
             with self.container(config=container_config) as container:
                 container.run(["/bin/touch", "/media/workdir/test"])
                 container.run(["/bin/test", "-e", "/media/workdir/test"])
                 self.assertTrue(os.path.exists(os.path.join(workdir, "test")))
 
                 binds = list(container.binds())
-                self.assertEqual(len(binds), 1)
+                self.assertIn(len(binds), (1, 2))
                 self.assertEqual(binds[0].source, workdir)
                 self.assertEqual(binds[0].destination, "/media/workdir")
                 self.assertEqual(binds[0].bind_type, "rw")
                 self.assertEqual(binds[0].mount_options, [])
+                if len(binds) == 2:
+                    # self.assertEqual(binds[1].source, workdir)
+                    self.assertEqual(binds[1].destination, "/var/cache/apt/archives")
+                    self.assertEqual(binds[1].bind_type, "rw")
+                    self.assertEqual(binds[1].mount_options, [])
 
     def test_bind_mount_ro(self):
         with tempfile.TemporaryDirectory() as workdir:
             # By default, things are run as root
             container_config = ContainerConfig()
-            container_config.binds.append(BindConfig(source=workdir, destination="/media/workdir", bind_type="ro"))
+            container_config.binds.append(
+                    BindConfig.create(source=workdir, destination="/media/workdir", bind_type="ro"))
             with self.container(config=container_config) as container:
                 res = container.run(["/bin/touch", "/media/workdir/test"], config=RunConfig(check=False))
                 self.assertEqual(res.returncode, 1)
@@ -214,29 +227,39 @@ class RunTestCase:
                 self.assertFalse(os.path.exists(os.path.join(workdir, "test")))
 
                 binds = list(container.binds())
-                self.assertEqual(len(binds), 1)
+                self.assertIn(len(binds), (1, 2))
                 self.assertEqual(binds[0].source, workdir)
                 self.assertEqual(binds[0].destination, "/media/workdir")
                 self.assertEqual(binds[0].bind_type, "ro")
                 self.assertEqual(binds[0].mount_options, [])
+                if len(binds) == 2:
+                    # self.assertEqual(binds[1].source, workdir)
+                    self.assertEqual(binds[1].destination, "/var/cache/apt/archives")
+                    self.assertEqual(binds[1].bind_type, "rw")
+                    self.assertEqual(binds[1].mount_options, [])
 
     def test_bind_mount_volatile(self):
         with tempfile.TemporaryDirectory() as workdir:
             # By default, things are run as root
             container_config = ContainerConfig()
             container_config.binds.append(
-                    BindConfig(source=workdir, destination="/media/workdir", bind_type="volatile"))
+                    BindConfig.create(source=workdir, destination="/media/workdir", bind_type="volatile"))
             with self.container(config=container_config) as container:
                 container.run(["/bin/touch", "/media/workdir/test"])
                 container.run(["/bin/test", "-e", "/media/workdir/test"])
                 self.assertFalse(os.path.exists(os.path.join(workdir, "test")))
 
                 binds = list(container.binds())
-                self.assertEqual(len(binds), 1)
+                self.assertIn(len(binds), (1, 2))
                 self.assertEqual(binds[0].source, workdir)
                 self.assertEqual(binds[0].destination, "/media/workdir")
-                self.assertEqual(binds[0].bind_type, "volatile")
+                self.assertEqual(binds[0].bind_type, "ro")
                 self.assertEqual(binds[0].mount_options, [])
+                if len(binds) == 2:
+                    # self.assertEqual(binds[1].source, workdir)
+                    self.assertEqual(binds[1].destination, "/var/cache/apt/archives")
+                    self.assertEqual(binds[1].bind_type, "rw")
+                    self.assertEqual(binds[1].mount_options, [])
 
     def test_run_callable_logging(self):
         def test_log():